requirements.txt
rns>=0.7.0
lxml>=4.9.0
requests>=2.31.0
pyyaml>=6.0
